23moped-26: Added exceptions to Repositorys - when constraints are null or they dont exist in the DB so no reference can be made - when id's for getter are < 0

+++ MopedMeetup/src/main/java/at/htl/meetup/controller/UserRepository.java
@@ -66,6 +66,9 @@ public class UserRepository {
     }
 
     public void delete(long id) {
+        if (id < 0) {
+            throw new IllegalArgumentException("ID must not be negative");
+        }
         try (Connection connection = dataSource.getConnection()) {
             String sql = "DELETE FROM MM_USER WHERE U_ID=?";
 
@@ -107,6 +110,9 @@ public class UserRepository {
     }
 
     public static User getById(long id){
+        if (id < 0) {
+            throw new IllegalArgumentException("ID must not be negative");
+        }
         try (Connection connection = dataSource.getConnection()) {
             String sql = "SELECT * FROM MM_USER WHERE U_ID=?";
             PreparedStatement statement = connection.prepareStatement(sql);
